adjust naming, format and scalafmt.conf as parameter
