+++ core/src/main/scala/caliban/parsing/adt/ExecutableDefinition.scala
@@ -6,8 +6,6 @@ sealed trait ExecutableDefinition
 
 object ExecutableDefinition {
 
-
-
   case class OperationDefinition(
     operationType: OperationType,
     name: Option[String],
