+++ core/src/main/scala/caliban/parsing/adt/Type.scala
@@ -11,7 +11,13 @@ object Type {
   case class NamedType(name: String, nonNull: Boolean) extends Type
   case class ListType(ofType: Type, nonNull: Boolean)  extends Type
 
-  case class FieldDefinition(description: Option[String], name: String, args: List[(String, Type)], ofType: Type, directives: List[Directive])
+  case class FieldDefinition(
+    description: Option[String],
+    name: String,
+    args: List[(String, Type)],
+    ofType: Type,
+    directives: List[Directive]
+  )
 
   @tailrec
   def innerType(t: Type): String = t match {
