+++ core/src/test/scala/caliban/parsing/ParserSpec.scala
@@ -5,10 +5,10 @@ import caliban.InputValue
 import caliban.InputValue._
 import caliban.Value._
 import caliban.parsing.ParserSpecUtils._
-import caliban.parsing.adt.ExecutableDefinition.{FragmentDefinition, OperationDefinition, TypeDefinition}
-import caliban.parsing.adt.OperationType.{Mutation, Query}
-import caliban.parsing.adt.Selection.{Field, FragmentSpread, InlineFragment}
-import caliban.parsing.adt.Type.{FieldDefinition, ListType, NamedType}
+import caliban.parsing.adt.ExecutableDefinition.{ FragmentDefinition, OperationDefinition, TypeDefinition }
+import caliban.parsing.adt.OperationType.{ Mutation, Query }
+import caliban.parsing.adt.Selection.{ Field, FragmentSpread, InlineFragment }
+import caliban.parsing.adt.Type.{ FieldDefinition, ListType, NamedType }
 import caliban.parsing.adt.{ Directive, Document, Selection, VariableDefinition }
 import zio.test.Assertion._
 import zio.test._
@@ -419,25 +419,32 @@ object ParserSpec
               |}""".stripMargin
           //planetSavedTimesCmp(moreThan: Int!)!: Int!
           assertM(
-            Parser.parseQuery(gqltype), equalTo(
+            Parser.parseQuery(gqltype),
+            equalTo(
               Document(
                 List(
-                  TypeDefinition("Hero",
+                  TypeDefinition(
+                    "Hero",
                     List(
-                      FieldDefinition(Some("name desc"), "name", List("pad" -> NamedType("Int", true)), NamedType("String", true), List(Directive("skip", Map("if" -> VariableValue("someTestM"))))),
+                      FieldDefinition(
+                        Some("name desc"),
+                        "name",
+                        List("pad" -> NamedType("Int", true)),
+                        NamedType("String", true),
+                        List(Directive("skip", Map("if" -> VariableValue("someTestM"))))
+                      ),
                       FieldDefinition(Some("nick desc"), "nick", List(), NamedType("String", true), List()),
                       FieldDefinition(None, "bday", List(), NamedType("Int", false), List()),
                       FieldDefinition(None, "suits", List(), ListType(NamedType("String", false), false), List()),
-                      FieldDefinition(None, "powers", List(), ListType(NamedType("String", true), true), List()),
+                      FieldDefinition(None, "powers", List(), ListType(NamedType("String", true), true), List())
 //                      FieldDefinition(None, "planetSavedTimesCmp", List("moreThan" -> NamedType("Int", true)), NamedType("Int", true), List()),
                     )
-
                   )
                 )
               )
             )
           )
-        },
+        }
       )
     )
