+++ core/src/test/scala/caliban/validation/ValidationSpec.scala
@@ -264,20 +264,20 @@ object ValidationSpec
           check(query, "Directive 'skip' is defined twice.")
         },
         testM("type has duplicate fields") {
-          val gqltype = gqldoc(
-            """type Hero {
+          val gqltype = gqldoc("""type Hero {
               name(pad: Int!): String! @skip(if: $someTestM)
               name: String!
               bday: Int
               }""")
           assertM(
             interpreter.check(gqltype).map(_.toString).run,
-            fails[CalibanError](hasField[CalibanError, String]("msg", _.msg, equalTo("Type 'Hero' has duplicate fields.")))
+            fails[CalibanError](
+              hasField[CalibanError, String]("msg", _.msg, equalTo("Type 'Hero' has duplicate fields."))
+            )
           )
         },
         testM("type defined twice") {
-          val gqltype = gqldoc(
-            """type Hero {
+          val gqltype = gqldoc("""type Hero {
                 name(pad: Int!): String! @skip(if: $someTestM)
                 nick: String!
                 bday: Int
@@ -290,7 +290,9 @@ object ValidationSpec
             """)
           assertM(
             interpreter.check(gqltype).map(_.toString).run,
-            fails[CalibanError](hasField[CalibanError, String]("msg", _.msg, equalTo("Type 'Hero' is defined more than once.")))
+            fails[CalibanError](
+              hasField[CalibanError, String]("msg", _.msg, equalTo("Type 'Hero' is defined more than once."))
+            )
           )
         }
       )
